pyproject.toml
[tool.poetry]
name = "caption"
version = "0.1.0"
description = ""
authors = ["sheldon <developer@8cylinder.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^9.4.0"
torch = "^1.13.1"
torchvision = "^0.14.1"
transformers = "^4.26.1"
requests = "^2.28.2"
click = "^8.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
